+++ classes/mage.lua
@@ -14,6 +14,7 @@ local UnitHealth = UnitHealth
 local clearcastingVariants; -- Lazy init in lazyCreateClearcastingVariants()
 
 local arcaneMissiles = 5143;
+local fireBlast = 2136;
 local pyroblast = 11366; -- Pyroblast, the base Pyro spell
 local pyroblastBang = 92315; -- Pyroblast!, a specific spell for instant Pyro introduced in Cataclysm
 
@@ -528,6 +529,26 @@ local function lazyCreateClearcastingVariants(self)
     });
 end
 
+local function useImpact()
+    local impactBuff = 64343;
+    local impactTalent = 11103;
+
+    SAO:CreateEffect(
+        "impact",
+        SAO.WRATH + SAO.CATA,
+        impactBuff,
+        "aura",
+        {
+            talent = impactTalent,
+            requireTalent = true,
+            action = fireBlast,
+            actionUsable = true,
+            overlay = { texture = "lock_and_load", position = "Top" },
+            button = fireBlast,
+        }
+    );
+end
+
 local function useArcaneMissiles()
     local arcaneMissilesBuff = 79683; -- Cataclysm requires a buff before casting Arcane Missiles
 
@@ -556,7 +577,7 @@ end
 
 local function registerClass(self)
     -- Fire Procs
-    self:RegisterAura("impact", 0, 64343, "lock_and_load", "Top", 1, 255, 255, 255, true, { (GetSpellInfo(2136)) });
+    useImpact();
     if self.IsWrath() then
         self:RegisterAura("firestarter", 0, 54741, "impact", "Top", 0.8, 255, 255, 255, true, { (GetSpellInfo(2120)) }); -- May conflict with Impact location
     end
@@ -672,9 +693,6 @@ local function loadOptions(self)
     local firestarterBuff = 54741;
     local firestarterTalent = 44442;
 
-    local impactBuff = 64343;
-    local impactTalent = 11103;
-
     local brainFreezeBuff = 57761;
     local brainFreezeTalent = 44546;
     local brainFreezeSoDRune = 400731;
@@ -692,7 +710,6 @@ local function loadOptions(self)
 
     local arcaneExplosion = 1449;
     local flamestrike = 2120;
-    local fireBlast = 2136;
     local fireball = 133;
     local frostfireBolt = 44614;
     local frostfireBoltSoD = 401502;
@@ -740,7 +757,6 @@ local function loadOptions(self)
         self:AddOverlayOption(hotStreakTalent, hotStreakHeatingUpBuff, 0, hotStreakHeatingUpDetails);
     end
     self:AddOverlayOption(firestarterTalent, firestarterBuff);
-    self:AddOverlayOption(impactTalent, impactBuff);
     if self.IsSoD() then
         self:AddOverlayOption(fingersOfFrostSoDTalent, fingersOfFrostSoDBuff, 0, nil, nil, 2); -- setup any stacks, test with 2 stacks
     elseif self.IsWrath() then
@@ -773,9 +789,6 @@ local function loadOptions(self)
     end
     if self.IsWrath() then
         self:AddGlowingOption(firestarterTalent, firestarterBuff, flamestrike);
-    end
-    if not self.IsEra() then -- Must exclude this option specifically for Classic Era, because the talent exists in Era but the proc is passive
-        self:AddGlowingOption(impactTalent, impactBuff, fireBlast);
     end
     if self.IsSoD() then
         self:AddGlowingOption(brainFreezeSoDRune, brainFreezeSoDBuff, fireball);